% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_antal_na.R
\name{kolonner_med_na}
\alias{kolonner_med_na}
\title{Antal na pr. kolonner
Funktionen tager en tabel og tæller NA'er i hver kolonne}
\usage{
kolonner_med_na(tabel)
}
\arguments{
\item{tabel}{bare en tabel}
}
\value{
en navngivet numerisk vector
}
\description{
Antal na pr. kolonner
Funktionen tager en tabel og tæller NA'er i hver kolonne
}
\examples{
x <- data.frame("tal"=c(1,2,3), "bogstaver"= c(NA, "b", "c"))
kolonner_med_na(x)
}
