% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_string_length.R
\name{kolonner_string_length}
\alias{kolonner_string_length}
\title{lav en tabel der viser antal for hver string_length pr. kolonne}
\usage{
kolonner_string_length(tabel)
}
\arguments{
\item{tabel}{en tabel med kolonner}
}
\value{
en tabel med kolonner og antal for hver string-length
}
\description{
lav en tabel der viser antal for hver string_length pr. kolonne
}
