% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_names}
\alias{sort_names}
\title{Sorter kolonnenavne som er tal
funktionen kan tage en tabel og sortere kolonnenavne som er tal i rette order, selvom tallet jo er en tekststreng}
\usage{
sort_names(data)
}
\arguments{
\item{data}{en tabel}
}
\value{
tabel med kolonner sorteret
}
\description{
Sorter kolonnenavne som er tal
funktionen kan tage en tabel og sortere kolonnenavne som er tal i rette order, selvom tallet jo er en tekststreng
}
