% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testen.R
\name{kolonner_med_na}
\alias{kolonner_med_na}
\title{Antal na pr. kolonner}
\usage{
kolonner_med_na(tabel)
}
\arguments{
\item{tabel}{}
}
\value{
en navngivet liste
}
\description{
Antal na pr. kolonner
}
\examples{
x <- data.frame("tal"=c(1,2,3), "bogstaver"= c(NA, "b", "c"))
kolonner_med_na(x)
}
